Moved settings to separate file, implemented settings, added insert feed option to allow viewing whole file.
