+++ settings.json
@@ -1,1 +1,1 @@
-{"dark_mode": false, "insert_timestamps": true, "add_horizontal_rule": true, "app_position": "top right", "win_width": 400, "win_height": 300, "font": "Courier", "font_size": 24, "hr_char": "-"}
+{"dark_mode": true, "insert_timestamps": true, "add_horizontal_rule": true, "app_position": "bottom right", "win_width": 450, "win_height": 300, "font_choice": "Courier", "font_size": 11, "hr_char": "-", "hotkey": "ctrl+alt+a", "word_wrap": true}
